+++ HISTORY.rst
@@ -169,6 +169,9 @@
 - Add document type change API endpoint.
 - Change OCR API submit URL from documents/{pk}/submit
   to documents/{pk}/ocr/submit.
+- Add Redis based distributed lock backend. Requires one
+  argument: "redis_url". Example: redis://127.0.0.1:6379/0
+- Add the setting LOCK_MANAGER_BACKEND_ARGUMENTS.
 
 3.3.11 (2019-XX-XX)
 ===================